+++ x/collection/keeper/supply.go
@@ -21,7 +21,7 @@ func (k Keeper) CreateContract(ctx sdk.Context, creator sdk.AccAddress, contract
 		panic(err)
 	}
 
-	eventGrant := collection.EventGrant{
+	eventGrant := collection.EventGranted{
 		ContractId: contractID,
 		Grantee:    creator.String(),
 	}
@@ -412,7 +412,7 @@ func (k Keeper) ModifyNFT(ctx sdk.Context, contractID string, tokenID string, op
 func (k Keeper) Grant(ctx sdk.Context, contractID string, granter, grantee sdk.AccAddress, permission collection.Permission) {
 	k.grant(ctx, contractID, grantee, permission)
 
-	event := collection.EventGrant{
+	event := collection.EventGranted{
 		ContractId: contractID,
 		Granter:    granter.String(),
 		Grantee:    grantee.String(),
@@ -433,7 +433,7 @@ func (k Keeper) grant(ctx sdk.Context, contractID string, grantee sdk.AccAddress
 func (k Keeper) Abandon(ctx sdk.Context, contractID string, grantee sdk.AccAddress, permission collection.Permission) {
 	k.deleteGrant(ctx, contractID, grantee, permission)
 
-	event := collection.EventAbandon{
+	event := collection.EventRenounced{
 		ContractId: contractID,
 		Grantee:    grantee.String(),
 		Permission: permission,